+++ R/sync.R
@@ -26,6 +26,9 @@
 #'     For \strong{wait} and \strong{until}: (invisibly) logical TRUE, or else
 #'     FALSE if a flag has been set.
 #'
+#'     For \strong{.until}: (invisibly) logical TRUE if signalled, or else FALSE
+#'     if the timeout was reached.
+#'
 #'     For \strong{cv_value}: integer value of the condition variable.
 #'
 #'     For \strong{cv_reset} and \strong{cv_signal}: zero (invisibly).
@@ -60,8 +63,8 @@
 #'
 #'     The condition variable also contains a flag that certain signalling
 #'     functions such as \code{\link{pipe_notify}} can set. When this flag has
-#'     been set, all subsequent \code{wait} or \code{until} calls will return
-#'     logical FALSE instead of TRUE.
+#'     been set, all subsequent \code{wait} calls will return logical FALSE
+#'     instead of TRUE.
 #'
 #'     Note that the flag is not automatically reset, but may be reset manually
 #'     using \code{cv_reset}.
@@ -76,7 +79,7 @@ cv <- function() .Call(rnng_cv_alloc)
 #' Condition Variables - Wait
 #'
 #' \code{wait} waits on a condition being signalled by completion of an
-#'     asynchronous receive.
+#'     asynchronous receive or pipe event.
 #'
 #' @param cv a 'conditionVariable' object.
 #'
@@ -91,7 +94,7 @@ wait <- function(cv) invisible(.Call(rnng_cv_wait, cv))
 #' Condition Variables - Until
 #'
 #' \code{until} waits until a future time on a condition being signalled by
-#'     completion of an asynchronous receive.
+#'     completion of an asynchronous receive or pipe event.
 #'
 #' @param msec maximum time in milliseconds to wait for the condition variable
 #'     to be signalled.
@@ -103,6 +106,11 @@ wait <- function(cv) invisible(.Call(rnng_cv_wait, cv))
 #' @export
 #'
 until <- function(cv, msec) invisible(.Call(rnng_cv_until, cv, msec))
+
+#' @rdname cv
+#' @export
+#'
+.until <- function(cv, msec) invisible(.Call(rnng_cv_until2, cv, msec))
 
 #' Condition Variables - Value
 #'
@@ -149,16 +157,17 @@ cv_signal <- function(cv) invisible(.Call(rnng_cv_signal, cv))
 #'
 #' @param socket a Socket.
 #' @param cv a 'conditionVariable' to signal.
-#' @param cv2 [default NULL] optionally, if specified, a second 'conditionVariable'
-#'     to signal. Note that this cv is signalled sequentially after the first
-#'     condition variable.
-#' @param add [default TRUE] logical value whether to signal when a pipe is added.
+#' @param cv2 [default NULL] optionally, if specified, a second
+#'     'conditionVariable' to signal. Note that this cv is signalled
+#'     sequentially after the first condition variable.
+#' @param add [default TRUE] logical value whether to signal when a pipe is
+#'     added.
 #' @param remove [default TRUE] logical value whether to signal when a pipe is
 #'     removed.
 #' @param flag [default TRUE] logical value whether to also set a flag in the
 #'     'conditionVariable'. This can help distinguish between different types of
-#'     signal, and causes any subsequent \code{\link{wait}} or \code{\link{until}}
-#'     to return FALSE instead of TRUE.
+#'     signal, and causes any subsequent \code{\link{wait}} to return FALSE
+#'     instead of TRUE.
 #'
 #' @details For add: this event occurs after the pipe is fully added to the
 #'     socket. Prior to this time, it is not possible to communicate over the
@@ -198,11 +207,16 @@ pipe_notify <- function(socket, cv, cv2 = NULL, add = TRUE, remove = TRUE, flag 
 
 #' Lock / Unlock a Socket
 #'
-#' Prevents further pipe connections from being established at a Socket.
+#' Prevents further pipe connections from being established at a Socket. If a
+#'     socket is locked, new pipe connections are closed before they can be
+#'     added to the socket.
 #'
 #' @param socket a Socket.
 #' @param cv (optional) a 'conditionVariable'. If supplied, the socket is locked
-#'     only while the value of the condition variable is non-zero.
+#'     only whilst the condition variable is an odd value. This is designed to
+#'     allow an initial connection, as well as subsequent re-connections after a
+#'     connection has ended, if the conditon variable is also registered with
+#'     \code{\link{pipe_notify}} for both add and remove pipe events.
 #'
 #' @return Invisibly, zero on success (will otherwise error).
 #'
@@ -235,63 +249,3 @@ lock <- function(socket, cv = NULL) invisible(.Call(rnng_socket_lock, socket, cv
 #' @export
 #'
 unlock <- function(socket) invisible(.Call(rnng_socket_unlock, socket))
-
-# nanonext - Weak References ---------------------------------------------------
-
-#' Weak References
-#'
-#' \code{weakref} creates a new weak reference - a special type of R object that
-#'     associates a Value with a Key. Value is kept alive for as long as Key
-#'     remains reachable (i.e. has yet to be garbage collected), even if Value
-#'     itself no longer has any references.
-#'
-#' @param key a reference object (such as an environment or external pointer).
-#' @param value an object.
-#'
-#' @return For \code{weakref}: a weak reference.
-#'
-#'     For \code{weakref_key} and \code{weakref_value}: the key or value
-#'     associated with the weak reference, or NULL if no longer reachable.
-#'
-#' @examples
-#' k <- new.env()
-#' v <- "value"
-#'
-#' w <- weakref(k, v)
-#' w
-#' typeof(w)
-#'
-#' @export
-#'
-weakref <- function(key, value) .Call(rnng_weakref_make, key, value)
-
-#' Weakref Key
-#'
-#' \code{weakref_key} retrieves the key associated with a weak reference.
-#'
-#' @param w a weak reference.
-#'
-#' @examples
-#' key <- weakref_key(w)
-#' identical(key, k)
-#'
-#' @rdname weakref
-#' @export
-#'
-weakref_key <- function(w) .Call(rnng_weakref_key, w)
-
-#' Weakref Value
-#'
-#' \code{weakref_value} retrieves the value associated with a weak reference.
-#'
-#' @examples
-#' value <- weakref_value(w)
-#' identical(value, v)
-#'
-#' rm(v)
-#' weakref_value(w)
-#'
-#' @rdname weakref
-#' @export
-#'
-weakref_value <- function(w) .Call(rnng_weakref_value, w)